Removed all warnings and changed a lot of the places we were using size_t to unsigned int instead. Started on light buffer

+++ gpupro-6/DataStructures/Mesh.cpp
@@ -60,7 +60,7 @@ const std::vector<UINT16>& Mesh::GetIndices()
 
 }
 
-size_t CreateMesh()
+int CreateMesh()
 {
 	Mesh* newMesh = new Mesh();
 	return MaterialManagementSystem::Instance()->RegisterMesh(*newMesh);
